move get_message_exception to utils
