+++ test/request_handler_test.py
@@ -6,6 +6,7 @@ from test_base_handler import TestBaseHandler
 from models.user import User
 from models.institution import Institution
 from models.request_user import RequestUser
+from utils import get_message_exception
 from handlers.request_handler import RequestHandler
 
 from mock import patch
@@ -170,10 +171,3 @@ def initModels(cls):
 
     cls.request = RequestUser.create(data)
     cls.request.put()
-
-
-def get_message_exception(cls, exception):
-    """Return only message of string exception."""
-    cls.list_args = exception.split("\n")
-    cls.dict = eval(cls.list_args[1])
-    return cls.dict["msg"]
